modeldeki relationshipler düzeltildi ,default değerler eklendi

+++ Api/Routers/manager.py
@@ -1,14 +1,14 @@
-from fastapi import APIRouter,Depends, Request
+from fastapi import APIRouter,Depends, Request,HTTPException
 from flask import jsonify
 from sqlalchemy.orm import Session
 from datetime import datetime,timedelta
 from Lib.models import SessionLocal,Project,Job,Comment,Status
 import json
 
-
+#gün kontrolü için eklediğim tarihler
 Now = datetime.today() 
-
 Future=datetime.today() + timedelta(days=2)
+
 
 managerroute = APIRouter()
 
@@ -26,14 +26,21 @@ def get_db():
 async def project_list(req: Request, db: Session = Depends(get_db)):
     """ istenilen kullanıcının bilgilerini veren fonksiyon"""
     projects = db.query(Project).all()
-    return projects
+    if not projects:
+        raise HTTPException(status_code=404,detail="Kullanıcı bulunamadı")
+    else :
+        return projects
+    
 
 
 @managerroute.get("/projects/{id}")
 async def get_project(req: Request,id:int, db: Session = Depends(get_db)):
     """ istenilen kullanıcının bilgilerini veren fonksiyon"""
     project = db.query(Project).filter_by(id=id).first()
-    return project
+    if project != None:
+        return project
+    else :
+        raise HTTPException(status_code=404,detail="Aranan Kullanıcı yoktur")
 
 
 @managerroute.post("/projects/add")
@@ -44,7 +51,7 @@ async def add_project(req: Request,db: Session = Depends(get_db)):
      project = Project( name= name)
      db.add(project)
      db.commit()
-     return print(name)
+     
 
 
 
@@ -55,19 +62,28 @@ async def update_project(req: Request,id:int,db: Session = Depends(get_db)):
    req_info = await  req.json()
    name = req_info['name']
    status= req_info['status']
-   db.query(Project).filter_by(id=id).update(
-    dict(name= name, status= status))
-   db.commit()
-   return "item updated"
+   project = db.query(Project).filter_by(id=id).first()
+   if project != None:
+     db.query(Project).filter_by(id=id).update(
+     dict(name= name, status= status))
+     db.commit()
+     return "proje güncellendi"
+   else :
+        raise HTTPException(status_code=404,detail="Aranan proje yoktur")
+  
 
 
 @managerroute.delete("/projects/delete/{id}")
 async def del_project(req: Request,id:int,db: Session = Depends(get_db)):
      """ kullanıcınun bilgilerini silen  fonksiyon """
      project = db.query(Project).filter_by(id=id).first()
-     db.delete(project)
-     db.commit()
-     return "veri silindi"
+     if project != None:
+       db.delete(project)
+       db.commit()
+       return "veri silindi"
+     else :
+       raise HTTPException(status_code=404,detail="Aranan proje yoktur")
+     
 
 #PROJELERİN APİ ÜZERİNDEN  CRUD REQUESTLERİ  BİTİŞ
 
@@ -77,14 +93,20 @@ async def del_project(req: Request,id:int,db: Session = Depends(get_db)):
 async def job_list(req: Request, db: Session = Depends(get_db)):
     """ istenilen kullanıcının bilgilerini veren fonksiyon"""
     jobs = db.query(Job).all()
-    return jobs
+    if not jobs:
+        raise HTTPException(status_code=404,detail="İş bulunamadı")
+    else :
+        return jobs
 
 
 @managerroute.get("/jobs/{id}")
 async def get_job(req: Request,id:int, db: Session = Depends(get_db)):
     """ istenilen kullanıcının bilgilerini veren fonksiyon"""
     job = db.query(Job).filter_by(id=id).first()
-    return job
+    if job != None:
+         return job
+    else :
+        raise HTTPException(status_code=404,detail="Aranan iş yoktur")
 
 
 @managerroute.post("/jobs/add")
@@ -98,7 +120,7 @@ async def add_job(req: Request,db: Session = Depends(get_db)):
      job = Job( title=titlee,content= content, status= status,project_id=project_id,created_at=Now, finish_date=Future)
      db.add(job)
      db.commit()
-     return print(content,status)
+    
 
 
 
@@ -110,19 +132,29 @@ async def update_job(req: Request,id:int,db: Session = Depends(get_db)):
    title = req_info['title']
    content = req_info['content']
    status= req_info['status']
-   db.query(Job).filter_by(id=id).update(
+   job = db.query(Job).filter_by(id=id).first()
+   if job  != None:
+    db.query(Job).filter_by(id=id).update(
     dict(title=title,content=content, status= status,updated_at=Now))
-   db.commit()
-   return "item updated"
+    db.commit()
+    return "item updated"
+   else :
+        raise HTTPException(status_code=404,detail="Aranan iş yoktur")
+
+        
+ 
 
 
 @managerroute.delete("/jobs/delete/{id}")
 async def del_job(req: Request,id:int,db: Session = Depends(get_db)):
-     """ kullanıcınun bilgilerini silen  fonksiyon """
-     job = db.query(Job).filter_by(id=id).first()
-     db.delete(job)
-     db.commit()
-     return "veri silindi"
+    """ kullanıcınun bilgilerini silen  fonksiyon """
+    job = db.query(Job).filter_by(id=id).first()
+    if job != None:
+      db.delete(job)
+      db.commit()
+      return "veri silindi"
+    else :
+        raise HTTPException(status_code=404,detail="Aranan iş yoktur")
 
 #JOBS(işlerin) APİ ÜZERİNDEN  CRUD REQUESTLERİ  BİTİŞ
 
@@ -131,15 +163,21 @@ async def del_job(req: Request,id:int,db: Session = Depends(get_db)):
 @managerroute.get('/comments')
 async def comment_list(req: Request, db: Session = Depends(get_db)):
     """ istenilen kullanıcının bilgilerini veren fonksiyon"""
-    jobs = db.query(Comment).all()
-    return jobs
+    comments = db.query(Comment).all()
+    if not comments:
+        raise HTTPException(status_code=404,detail="Yorum bulunamadı")
+    else :
+        return comments
 
 
 @managerroute.get("/comments/{id}")
 async def get_comment(req: Request,id:int, db: Session = Depends(get_db)):
     """ istenilen kullanıcının bilgilerini veren fonksiyon"""
     comment = db.query(Comment).filter_by(id=id).first()
-    return comment
+    if comment  != None:
+        return comment 
+    else :
+        raise HTTPException(status_code=404,detail="Aranan yorum yoktur")
 
 
 
@@ -152,7 +190,7 @@ async def add_commentt(req: Request,db: Session = Depends(get_db)):
      comment = Comment( comment=commentt,job_id=job_id)
      db.add(comment)
      db.commit()
-     return print(comment)
+     
 
 
 
@@ -163,20 +201,33 @@ async def update_comment(req: Request,id:int,db: Session = Depends(get_db)):
    req_info = await  req.json()
    comment = req_info['comment'] 
    job_id=int(req_info['job_id'])
-   db.query(Comment).filter_by(id=id).update(
-    dict(comment=comment,job_id=job_id))
-   db.commit()
-   return "item updated"
+   commentt = db.query(Comment).filter_by(id=id).first()
+   if commentt!= None:
+       db.query(Comment).filter_by(id=id).update(
+       dict(comment=comment,job_id=job_id))
+       db.commit()
+       return "item updated"
+   else :
+        raise HTTPException(status_code=404,detail="Aranan Kullanıcı yoktur")
+
+   
+
+
+
+   
 
 
 
 @managerroute.delete("/comments/delete/{id}")
 async def del_comment(req: Request,id:int,db: Session = Depends(get_db)):
-     """ kullanıcınun bilgilerini silen  fonksiyon """
+     """ yorum silen  fonksiyon """
      comment = db.query(Comment).filter_by(id=id).first()
-     db.delete(comment)
-     db.commit()
-     return "veri silindi"
+     if comment != None:
+      db.delete(comment)
+      db.commit()
+      return "yorum silindi"
+     else :
+        raise HTTPException(status_code=404,detail="Aranan Yorum yoktur")
 
 #Comments (yorumların) APİ ÜZERİNDEN  CRUD REQUESTLERİ  BİTİŞ
 
@@ -184,15 +235,21 @@ async def del_comment(req: Request,id:int,db: Session = Depends(get_db)):
 @managerroute.get('/status')
 async def status_list(req: Request, db: Session = Depends(get_db)):
     """ istenilen kullanıcının bilgilerini veren fonksiyon"""
-    jobs = db.query(Status).all()
-    return jobs
+    statuss = db.query(Status).all()
+    if not statuss:
+        raise HTTPException(status_code=404,detail="Statüler bulunamadı")
+    else :
+        return statuss
 
 
 @managerroute.get("/status/{id}")
 async def get_status(req: Request,id:int, db: Session = Depends(get_db)):
     """ istenilen kullanıcının bilgilerini veren fonksiyon"""
-    comment = db.query(Status).filter_by(id=id).first()
-    return comment
+    status = db.query(Status).filter_by(id=id).first()
+    if status != None:
+        return status
+    else :
+        raise HTTPException(status_code=404,detail="Aranan durum yoktur")
 
 
 
@@ -204,7 +261,7 @@ async def add_status(req: Request,db: Session = Depends(get_db)):
      status = Status( status_name=status_name)
      db.add(status)
      db.commit()
-     return print(status)
+     
 
 
 
@@ -213,11 +270,17 @@ async def add_status(req: Request,db: Session = Depends(get_db)):
 async def update_status(req: Request,id:int,db: Session = Depends(get_db)):
    """ kullanıcınun bilgilerini editleyen  fonksiyon """
    req_info = await  req.json()
-   status_name= req_info['status_name'] 
-   db.query(Status).filter_by(id=id).update(
-    dict(status_name=status_name))
-   db.commit()
-   return "item updated"
+   status_name= req_info['status_name']
+   status = db.query(Status).filter_by(id=id).first()
+   if status != None:
+      db.query(Status).filter_by(id=id).update(
+      dict(status_name=status_name))
+      db.commit()
+      return "item updated"
+   else :
+        raise HTTPException(status_code=404,detail="Aranan durum yoktur")
+
+  
 
 
 
@@ -225,8 +288,12 @@ async def update_status(req: Request,id:int,db: Session = Depends(get_db)):
 async def del_status(req: Request,id:int,db: Session = Depends(get_db)):
      """ kullanıcınun bilgilerini silen  fonksiyon """
      comment = db.query(Status).filter_by(id=id).first()
-     db.delete(comment)
-     db.commit()
-     return "veri silindi"
+     if comment != None:
+      db.delete(comment)
+      db.commit()
+      return "veri silindi"
+     else :
+        raise HTTPException(status_code=404,detail="Aranan Kullanıcı yoktur")
+
 
 #Status (durumlar) APİ ÜZERİNDEN  CRUD REQUESTLERİ BİTİŞ
